Split alternate roleset deletion component responder into separate class to make it non-ephemeral. Delete alternate roleset deletion message so after interaction is complete so it's not sitting around forever letting alt rolesets be deleted.
